Move this SSM code out to a different file
